researcher/pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "AI-powered deep research agent using Google Gemini for comprehensive knowledge synthesis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Jaime Cernuda" }
]
keywords = ["deep-research", "gemini", "ai-agent", "research", "knowledge-synthesis", "google-genai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI & Output
    "click>=8.1.0",
    "rich>=13.0.0",
    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # Google Gemini API
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
researcher = "researcher.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/researcher"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_return_any = false
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["yaml", "aiofiles"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "network: requires network access (run with --network)",
    "slow: marks tests as slow running",
]

[dependency-groups]
dev = []
